Implement SwitchTab for Nunit
