+++ common/src/codegen/playwright-charp-common/dotnetTemplateCollection.ts
@@ -1,5 +1,5 @@
-import { TemplateCollectionOptions } from "../playwright-csharp-mstest/templateCollection";
 import { BaseTemplateCollection } from "../codegen-common/baseTemplateCollection";
+import { TemplateCollectionOptions } from "./commonCsharpTemplateCollection";
 
 export class BaseDotnetTemplateCollection extends BaseTemplateCollection {
   public readonly ENVIRONMENT_SETTINGS_FILE: HandlebarsTemplateDelegate<any>;
