+++ common/src/codegen/playwright-charp-common/templates/TestRoutineBaseFile.hbs
@@ -1,4 +1,4 @@
-namespace {{{rootNamespace}}}.MsTestSupport;
+namespace {{{rootNamespace}}}.Support;
 
 /// <summary>
 /// Base class for RM test routine
