+++ common/src/codegen/playwright-csharp-mstest/playwrightCsharpMsTestTemplatesProvider.ts
@@ -41,4 +41,8 @@ export class PlaywrightCsharpMsTestTemplatesProvider {
   getRunSettings() {
     return this._templateCollection.RUNSETTINGS_FILE({});
   }
+
+  getTestFunction(name: string, description: string) {
+    return this._templateCollection.TEST_FUNCTION({ name, description });
+  }
 }
