+++ common/src/codegen/playwright-csharp-mstest/templateCollection.ts
@@ -1,10 +1,3 @@
-import { compile } from "handlebars";
-import { loadTemplate } from "../utils/templateLoader";
-import { ICsharpTemplateCollection } from "../playwright-charp-common/csharpTemplateCollection";
-import { BaseDotnetTemplateCollection } from "../playwright-charp-common/dotnetTemplateCollection";
-import { Indent } from "../../file-defs";
-import { EOL } from "os";
-import { indentCharMap } from "../../file-defs/shared";
 import { BaseTemplateCollection } from "../codegen-common/baseTemplateCollection";
 import { TemplateCollectionOptions } from "../playwright-charp-common/commonCsharpTemplateCollection";
 
@@ -16,6 +9,7 @@ export class MsTestTemplateCollection extends BaseTemplateCollection {
   public readonly CSPROJECT_FILE: HandlebarsTemplateDelegate<any>;
   public readonly RUNSETTINGS_FILE: HandlebarsTemplateDelegate<any>;
   public readonly USINGS_FILE: HandlebarsTemplateDelegate<any>;
+  public readonly TEST_FUNCTION: HandlebarsTemplateDelegate<any>;
 
   constructor(templateCollectionOptions: TemplateCollectionOptions) {
     super(templateCollectionOptions);
@@ -25,5 +19,7 @@ export class MsTestTemplateCollection extends BaseTemplateCollection {
     this.CSPROJECT_FILE = this.loadAndCompile("CsProjectFile");
     this.RUNSETTINGS_FILE = this.loadAndCompile("RunSettingsFile");
     this.USINGS_FILE = this.loadAndCompile("UsingsFile");
+
+    this.TEST_FUNCTION = this.loadAndCompile("TestFunction");
   }
 }
