+++ common/src/codegen/playwright-csharp-nunit/templateCollection.ts
@@ -1,16 +1,15 @@
-import { compile } from "handlebars";
 import { BaseTemplateCollection } from "../codegen-common/baseTemplateCollection";
 import { TemplateCollectionOptions } from "../playwright-charp-common/commonCsharpTemplateCollection";
 
 /** Templates collection for Playwright Csharp Nunit codegen*/
 export class NunitTemplateCollection extends BaseTemplateCollection {
-  
   public readonly TEST_SUITE_BASE_FILE: HandlebarsTemplateDelegate<any>;
   public readonly TEST_SUITE_FILE: HandlebarsTemplateDelegate<any>;
 
   public readonly CSPROJECT_FILE: HandlebarsTemplateDelegate<any>;
   public readonly RUNSETTINGS_FILE: HandlebarsTemplateDelegate<any>;
   public readonly USINGS_FILE: HandlebarsTemplateDelegate<any>;
+  public readonly TEST_FUNCTION: HandlebarsTemplateDelegate<any>;
 
   constructor(templateCollectionOptions: TemplateCollectionOptions) {
     super(templateCollectionOptions);
@@ -20,6 +19,6 @@ export class NunitTemplateCollection extends BaseTemplateCollection {
     this.CSPROJECT_FILE = this.loadAndCompile("CsProjectFile");
     this.RUNSETTINGS_FILE = this.loadAndCompile("RunSettingsFile");
     this.USINGS_FILE = this.loadAndCompile("UsingsFile");
-  }
+    this.TEST_FUNCTION = this.loadAndCompile("TestFunction");
   }
 }
