--- common/src/codegen/playwright-csharp-nunit/templates/BasePageFile.hbs
@@ -1,8 +1,0 @@
-namespace {{{rootNamespace}}};
-
-public abstract class PageBase
-{
-    protected IPage Page => this.PageTest.Page;
-    protected readonly BasePageTest PageTest;
-    public PageBase(BasePageTest pageTest) => PageTest = pageTest;
-}
