--- common/src/codegen/playwright-csharp-nunit/templates/Comment.hbs
@@ -1,1 +1,0 @@
-// {{{message}}}
