--- common/src/codegen/playwright-csharp-nunit/templates/PageDefinitionsFile.hbs
@@ -1,17 +1,0 @@
-{{{usings}}}
-
-namespace {{{rootNamespace}}};
-
-public class PageDefinitions
-{
-    public IPage Page { get; private set; }
-
-{{{pageDeclaration}}}
-
-    public PageDefinitions(IPage page)
-    {
-        Page = page;
-
-{{{body}}}
-    }
-}
