--- common/src/codegen/playwright-csharp-nunit/templates/PageElementProperty.hbs
@@ -1,9 +1,0 @@
-{{#if description}}
-/// <summary>
-/// {{{description}}}
-/// </summary>
-{{/if}}
-public {{{returnedLocatorType}}} {{{elementName}}}({{#if hasParams}}params string[] parameters{{/if}})
-{
-    return {{{getter}}};
-}
