--- common/src/codegen/playwright-csharp-nunit/templates/PageFile.hbs
@@ -1,19 +1,0 @@
-namespace {{{fullNamespace}}};
-
-{{#if pageDescription}}
-/// <summary>
-/// {{{pageDescription}}}
-/// </summary>
-{{/if}}
-public class {{{pageName}}}
-{
-    private IPage _page;
-
-    /// <summary>
-    /// Default constructor
-    /// </summary>
-    /// <param name="page">Playwright Page</param>
-    public {{{pageName}}}(IPage page) => _page = page;
-
-{{{pageBody}}}
-}
