--- common/src/codegen/playwright-csharp-nunit/templates/TestCaseBaseFile.hbs
@@ -1,35 +1,0 @@
-namespace {{{rootNamespace}}}.Support;
-
-/// <summary>
-/// Base class for RM "Single case" test suite 
-/// </summary>
-public abstract class TestCaseBase : PageTest
-{
-    protected PageTest? TestSuiteInstance;
-
-    /// <summary>
-    /// Returns TestSuiteInstance.Page when TestSuiteInstance is not null, otherwise return the Page of base class (PageTest.Page)
-    /// </summary>
-    public new IPage Page => TestSuiteInstance == null ? base.Page : TestSuiteInstance.Page;
-
-    /// <summary>
-    /// Default constructor
-    /// </summary>
-    public TestCaseBase() { }
-
-    /// <summary>
-    /// Create new instance of TestCaseBase with the provided "caller" testSuiteInstance
-    /// </summary>
-    public TestCaseBase(PageTest testSuiteInstance)
-    {
-        this.TestSuiteInstance = testSuiteInstance;
-    }
-    
-    /// <summary>
-    /// Custom Expect method that works from IFrameLocator
-    /// </summary>
-    public ILocatorAssertions Expect(IFrameLocator locator)
-    {
-        return base.Expect(locator.Locator(":root"));
-    }
-}
