--- common/src/codegen/playwright-csharp-nunit/templates/TestCaseFile.hbs
@@ -1,22 +1,0 @@
-{{#if usings.length}}
-{{#each usings as |using|}}
-{{{using}}}
-{{/each}}
-{{/if}}
-
-namespace {{{fullNamespace}}};
-
-/// <summary>
-/// {{{description}}}
-/// </summary>
-public class {{{testCaseName}}} : TestCaseBase
-{
-    public {{{testCaseName}}}(PageTest testSuiteInstance): base(testSuiteInstance) {}
- 
-    public async Task RunAsync()
-    {
-        PageDefinitions defs = new PageDefinitions(this.Page);
-
-{{{body}}}
-    }
-}
