+++ common/src/codegen/playwright-csharp-nunit/templates/TestFunction.hbs
@@ -4,5 +4,5 @@
 [Test]
 public async Task {{{name}}}()
 {
-    await new {{{name}}}(this).RunAsync();
+    await new {{{name}}}(this.Page).RunAsync();
 }
