--- common/src/codegen/playwright-csharp-nunit/templates/TestRoutineClass.hbs
@@ -1,14 +1,0 @@
-/// <summary>
-/// {{{description}}}
-/// </summary>
-public partial class {{{testRoutineName}}} : TestRoutineBase
-{
-    public {{{testRoutineName}}}(IPage page) : base(page) { }
-
-    public async Task RunAsync()
-    {
-        PageDefinitions defs = new PageDefinitions(this.Page);
-
-{{{body}}}
-    }
-}
