--- common/src/codegen/playwright-csharp-nunit/templates/TestRoutineFile.hbs
@@ -1,7 +1,0 @@
-using {{{rootNamespace}}}.TestRoutines;
-
-namespace {{{fullNamespace}}};
-
-{{#each testRoutineClasses as |testRoutineClass|}}
-{{{testRoutineClass}}}
-{{/each}}
