+++ common/src/codegen/playwright-csharp-xunit/playwrightCsharpXUnitCodeGen.ts
@@ -136,4 +136,13 @@ export class PlaywrightCsharpXUnitCodeGen extends CommonPlaywrightCsharpCodeGen 
     );
     return testClass;
   }
+
+  protected generateTestCaseFunction(testCase: ITestCase) {
+    const testcaseName = this._outProjMeta.get(testCase.id)!.outputFileClassName;
+    const testCaseMethod = this._templateProvider.getTestFunction(
+      upperCaseFirstChar(testcaseName),
+      testCase.description
+    );
+    return testCaseMethod;
+  }
 }
